moving one git related file into app/lib/git
